current state - stage four effectively complete, 3 mostly
we now have rendering in place.

altered the plot bitmap to XOR the bits, so white or black background,
it'll plot accordingly.

Also allowed a HUGE cheat with rendering, I can xor the last cycle
position to undraw only the cycle, then draw at the new position, a
whopping 32 byte redraw to deal with the cycles!

Collision and other events not fully in place, hence stage three
incomplete.  The tst_rast.c code had a little tacked on to call the
stage 4 rendering code, and loop moving.  It now places the cycles at
their respective starts, and sets them in motion and runs them through
eachother.  and works accordingly.

Have some constants prepped, not yet instituted.  Had to revert to last
state because I converted too many values in one go without testing,
broke it badly, lol.  Kept the constants for the next attempt.

New enum, velocity, allows setting the speed with slow, norm, and fast,
to assign 2, 4,  and 6 respectively.

model updated to allow tracking last position and fixed abug where I
assigned the wrong crasft bitmap

rewrote the stickman to be black on white, so that the XOR code woud
draw it accordingly on white or black background

huge rewrite of rndr_cyc: it now undraws the last position, and then
draws the new position.  planing to update it so that render cycle makes
the call to get the light wall drawn, since it knows where it was and
now is.